global Patch 241187_001001.ipf

+++ global/addon.ipf/warningmsgbox/warningmsgbox.lua
@@ -1,16 +1,20 @@
 ﻿-- warningmsgbox.lua
-
+local local_item_grade = nil
 
 function WARNINGMSGBOX_ON_INIT(addon, frame)
 	addon:RegisterMsg("DO_OPEN_WARNINGMSGBOX_UI", "WARNINGMSGBOX_FRAME_OPEN");
 end
 
-function WARNINGMSGBOX_FRAME_OPEN(clmsg, yesScp, noScp, itemGuid)
+function WARNINGMSGBOX_FRAME_OPEN(clmsg, yesScp, noScp, itemGuid)        
 	ui.OpenFrame("warningmsgbox")
 	
 	local frame = ui.GetFrame('warningmsgbox')
 	local warningText = GET_CHILD_RECURSIVELY(frame, "warningtext")
 	warningText:SetText(clmsg)
+
+    local input_frame = GET_CHILD_RECURSIVELY(frame, "input")
+    input_frame:ShowWindow(1)
+    input_frame:SetText('')
 
 	local showTooltipCheck = GET_CHILD_RECURSIVELY(frame, "cbox_showTooltip")
 	if itemGuid ~= nil then
@@ -20,9 +24,24 @@ function WARNINGMSGBOX_FRAME_OPEN(clmsg, yesScp, noScp, itemGuid)
 	else
 		showTooltipCheck:ShowWindow(0)
 	end
-
+    
+    if itemGuid ~= nil then
+        local item = session.GetInvItemByGuid(itemGuid)
+        if item ~= nil then
+            local_item_grade = GetIES(item:GetObject()).ItemGrade
+        else
+            local_item_grade = 0
+        end
+    else
+        local_item_grade = 0
+    end
+    
 	local yesBtn = GET_CHILD_RECURSIVELY(frame, "yes")
 	tolua.cast(yesBtn, "ui::CButton");
+
+    if local_item_grade < 3 then
+        input_frame:ShowWindow(0)            
+    end
 
 	yesBtn:SetEventScript(ui.LBUTTONUP, '_WARNINGMSGBOX_FRAME_OPEN_YES');
 	yesBtn:SetEventScriptArgString(ui.LBUTTONUP, yesScp);
@@ -31,18 +50,29 @@ function WARNINGMSGBOX_FRAME_OPEN(clmsg, yesScp, noScp, itemGuid)
 	tolua.cast(noBtn, "ui::CButton");
 
 	noBtn:SetEventScript(ui.LBUTTONUP, '_WARNINGMSGBOX_FRAME_OPEN_NO');
-	noBtn:SetEventScriptArgString(ui.LBUTTONUP, noScp);
+	noBtn:SetEventScriptArgString(ui.LBUTTONUP, noScp)
 
-	local buttonMargin = noBtn:GetMargin();
-	local warningbox = GET_CHILD_RECURSIVELY(frame, 'warningbox');
-	local totalHeight = warningbox:GetY() + warningText:GetY() + warningText:GetHeight() + showTooltipCheck:GetHeight() + noBtn:GetHeight() + 2 * buttonMargin.bottom;
-	local bg = GET_CHILD_RECURSIVELY(frame, 'bg');
-	warningbox:Resize(warningbox:GetWidth(), totalHeight);
-	bg:Resize(bg:GetWidth(), totalHeight);
-	frame:Resize(frame:GetWidth(), totalHeight);	
+	local buttonMargin = noBtn:GetMargin()
+	local warningbox = GET_CHILD_RECURSIVELY(frame, 'warningbox')
+	local totalHeight = warningbox:GetY() + warningText:GetY() + warningText:GetHeight() + showTooltipCheck:GetHeight() + noBtn:GetHeight() + 2 * buttonMargin.bottom + input_frame:GetHeight()
+    if itemGuid == nil or local_item_grade < 3 then
+        totalHeight = warningbox:GetY() + warningText:GetY() + warningText:GetHeight() + showTooltipCheck:GetHeight() + noBtn:GetHeight() + 2 * buttonMargin.bottom
+    end
+
+	local bg = GET_CHILD_RECURSIVELY(frame, 'bg')
+	warningbox:Resize(warningbox:GetWidth(), totalHeight)
+	bg:Resize(bg:GetWidth(), totalHeight)
+	frame:Resize(frame:GetWidth(), totalHeight)
 end
 
 function _WARNINGMSGBOX_FRAME_OPEN_YES(parent, ctrl, argStr, argNum)
+    local input_frame = GET_CHILD_RECURSIVELY(parent, "input")    
+    if local_item_grade >= 3 and input_frame:GetText() ~= '0000' then
+        -- 확인메시지 불일치
+        ui.SysMsg(ClMsg('miss_match_confirm_text'))
+        return
+    end
+
 	IMC_LOG("INFO_NORMAL", "_WARNINGMSGBOX_FRAME_OPEN_YES" .. argStr)
 	local scp = _G[argStr]
 	if scp ~= nil then
@@ -105,8 +135,15 @@ function WARNINGMSGBOX_CREATE_TOOLTIP(frame)
 
 	tooltipFrame:SetTooltipStrArg('inven');
 	tooltipFrame:SetTooltipIESID(itemGuid);
-    tooltipFrame:RefreshTooltip();
-	tooltipFrame:SetOffset(warningboxFrame:GetX() + warningboxFrame:GetWidth(), warningboxFrame:GetY())
+	tooltipFrame:RefreshTooltip();
+
+	-- 툴팁 출력위치 조정
+	local OffsetRatioM = frame:GetUserConfig("TOOLTIP_OFFSET_M");
+	local OffsetRatioS = frame:GetUserConfig("TOOLTIP_OFFSET_S");
+	local OffsetX = warningboxFrame:GetX() + warningboxFrame:GetWidth() - ( tooltipFrame:GetWidth() / OffsetRatioM );
+	local OffsetY = warningboxFrame:GetY() - ( tooltipFrame:GetHeight() / OffsetRatioS );
+	tooltipFrame:SetOffset(OffsetX, OffsetY)
+
 	local isShowTooltip = config.GetXMLConfig("ShowTooltipInWarningBox")
 	if isShowTooltip == 1 then
 		tooltipFrame:ShowWindow(1)
@@ -133,4 +170,7 @@ function WARNINGMSGBOX_SHOW_TOOLTIP(frame)
 	else
 		tooltipFrame:ShowWindow(0)
 	end
+end
+
+function UPDATE_TYPING_SCRIPT_WARNINGMSGBOX(frame, ctrl)	
 end
